Convert pgnodemx_memstat_int64 into pgnodemx_cgroup_scalar_bigint
First step into conversion to handle all controllers with a common
set of SQL facing functions. The cgroup virtual files all look like
<controller>.<metric>. Use that fact to parse the controller name
out of the filename. On cgroup v2 it does not actually matter since
all the controllers are unified under one cgroup. But on cgroup v1
it matters greatly since the cgroups are underneath each of the
controllers (and in fact independent even if named the same).

This makes use of the previous commits work to ensure cgpath
is populated with a path per controller.