Split cgroup functions into own source file
In preparation for adding non-cgroup functionality, split the cgroup
specific code into its own files. Some of what moved should probably
be further split out to a utility/common source file, but this will
do for now. This leaves only the init function and the SQL accessible
functions in the main source file.