Move cgroupfs_enabled check up from setof_scalar_internal
cgroupfs_enabled check was being performed at the setof_scalar_internal
level. However this is the wrong place to do that. It really belongs at
the SQL function level, because setof_scalar_internal is otherwise generic
and in a future commit will be moved out of cgroup.c entirely and into
parseutils.c.