Move and rename cgroup_setof_scalar_internal
Move cgroup_setof_scalar_internal to genutils.c and rename to
setof_scalar_internal. Also remove the ties to cgroup by moving
the calculation of the fully qualified path to the file of interest up
to the SQL function level. The function is now suitable for general use,
except that it still tries to do too much -- but that will be addressed
in the next commit.