Ensure CGROUP2_SUPER_MAGIC is defined, add proc_meminfo()
In RHEL 7 CGROUP2_SUPER_MAGIC is not defined in <linux/magic.h> because
there is not yet cgroup v2 support. The extension will detect that it is
operating under cgroup v1, but the define is needed just to compile in
the first place, so arrange to define it if the header does not.

Additionally added proc_meminfo() and do a fair bit of related
refactoring. In particular, modify and rename the underlying line
parsing function to make it useful for a more general case.