Properly handle path lookup when controller is multi-keyed
When the controller listed in /proc/self/cgroup is multi-keyed, for
example "cpu,cpuacct", and trying to look up e.g. "cpu.shares", the
lookup was failing. Fix that by detecting multi-keyed controller names
and checking for a match against each part.

Also clone sql/pgnodemx.sql to make sql/pgnodemx_2.sql, and create a
new sql/pgnodemx_1.sql. The "_1" is for cgroup v1, and "_2" is cgroup
v2. These are still not real regression tests, but stepping in that
general direction.

Note that the README.md is now a bit behind reality -- fix that in the
next commit.