Add note to README about role membership
Prior to PostgreSQL v10 no default pg_monitor role exists. The version
portability patch recently committed made it possible in pre-v10
PostgreSQL to check for membership in a special role named "pgmonitor"
rather than the pg10+ default role called pg_monitor. The former is
required because Postgres will not allow roles to be created that start
with "pg_". Note this compatibility issue in the README.