First pass at bugfix for cgroup "list" controller ordering
cgroup controllers under cgroup v1 are sometimes aggregated into
a "list" for the purposes of the Linux kernel. As an example, the
cpuacct and cpu controllers are aggregated together in a
"cpuacct,cpu" directory. These controllers and directories are also
referenced in the /proc/<pid>/cgroup files for processes running in
cgroups. Until now, the assumption was made that the ordering of
the controller names would be consistent between the proc file and
the actual cgroup directory names. However at least one instance in
the wild shows this to be a bad assumption.

Fix the ordering issue by first checking for the existence of the
directory based on the proc file spelling, and then falling back to
an exhaustive search of all permutations if that spelling does not
exist as the actual combined controller directory name.