Ensure the default cgroup controller path is initialized
It was theoretically possible that defpath could be used uninitialized,
which would obviously be a bad thing. Initialize defpath to NULL and
modify create_default_cgpath() so that if defpath is NULL, the path is
set up as "Default_Controller_Not_Found". This is done rather than
throwing an error because it seems friendlier and more helpful for
debugging since otherwise postgres would fail to start entirely.