Modify heap_permute and related to avoid lists
Refactor heap_permute and related functionality to deal with arrays of
char* and arrays of int rather than passing everything around as
postgres lists. Be more proactive about freeing memory along the way as
well.

Also restrict the list length of the input to 10, since
the number of permutations starts getting a bit unwieldy. In fact by 12
items in a list the number of permutations is almost 500 million, so
even an array of pointers starts to approach 4GB. For our use case here
I don't think we will ever need more than 3 or possibly four items in
the input list so we will go with 10 until/unless proven otherwise
somewhere down the road.