Add functions to detect FIPS mode & symbol source
Add new function to detect the "FIPS mode" under which PostgreSQL is
currently operating. Also add a function to provide the name of the
source dynamic library for a given C function by sysmbol name. This
latter function is useful to verify that the "FIPS_mode" call used by
the former function actually comes from the correct FIPS validated
library.