Fix install script version, add lib version func
PostgreSQL extensions have a version associated with the extension's
installation script, which is not always the same as the "version" of
the installed shared object library. Basically the install script
version only changes when the SQL interface for the library changes,
but other changes such as bug fixes which do not have SQL interface
changes associated with them are not captured by the install script
version.

The last released version of the library did not change the install
script and therefore the install script version did not change at that
time. Unfortunately that means in the field there is not always an
easy way to determine if the deplyed version of the library
definitively includes a particular bug fix. Correct that deficiency
by adding a new SQL function, pgnodemx_version(), which will display
a value based on one of three possible compile time scenarios:
* If the "VSTR" environment valiable is set at compile time, the
  function returns that value
* Otherwise it returns the value of the short git hash...
* Unless. if not compiling from the git repository *and* VSTR is unset,
  it returns "none"

The SQL install script *did* change with this commit, so bump the
install script version as well. Note that this fixes an infelicity in
the previous commit (which was post last release) since it also changed
the install script but the install script version was not bumped as it
should have been in that commit.