Fix Makefile, add role check to new FIPS functions
The last commit was a few bricks shy of a load as far as the Makefile
changes were concerned, so fix that.

Also, while at it, ensure that fips_mode() and symbol_filename() check
for pg_monitor role. These functions provide a bit too much
introspection for the casual user so let's at least start out by
restricting their use a bit.