Add pg_proctab equivalent functionality
Add procfs functions that expose the same basic information
as pg_proctab, but do it in a way consistent with the rest
of pgnodemx. Also provide wrapper compatibility functions
with the hope that pgnodemx can serve as a remote for pg_top.
So far this is untested.

This commit lacks proper documentation, but it is getting
unweildy and is feature complete I believe, so look for the
missing docs in a later commit.