Add exec_path() and fix bug/comment/pg version compatibility
Turns out <extension>.control file is needed even when a specific
version is being installed, and even if it is empty. So fix that
by adding a dummy file, however arrange things so that the dummy
file is only installed if another does not already exist. This will
hopefully prevent situations where the actual pg_proctab extension
is installed and pgnodemx is added later.

Additionally fix pgnodemx_pages_to_bytes() for pg versions before
pg12. Also fix some incorrect comments.

Finally add the first of a new series of functions intended to
allow introspection for compliance checking -- exec_path().

(adding the rest of the files -- botched the last commit)