Fully support proc diskstats 14, 18, or 20 fields
/proc/diskstats has 14 fields prior to kernel 4.18 and 18 fields prior
to kernel 5.5. The lines from that file were being parsed correctly and
checked for one of those lengths previously, however the last 6 fields
were being unconditionally discarded. Now they are preserved in the
output when present, and NULL values passed when they are not.

Also add these fields to the recently committed pg_proctab compatibility
function. In this case any NULL fields are COALESCE'd to 0, since that
is what pg_proctab itself does.