Update pg-proctab compat to 0.0.10
The previous commit modified the pg-proctab compat function
pg_diskusage() to be equivalent to the original pg-proctab 0.0.10
output. Namely, it adds 6 columns. Apparently the github mirror of
pg_proctab never got the memo, but the canonical gitlab repo is what we
should be measuring ourselves against. In any case, update all of the
supporting files and documentation to reflect the correct 0.0.10
version.